[darwin] move CoreAudioHelpers into darwin folder
