+++ xbmc/cores/AudioEngine/Sinks/darwin/CoreAudioHelpers.cpp
