+++ xbmc/cores/AudioEngine/Sinks/darwin/CoreAudioHelpers.h
