+++ xbmc/cores/AudioEngine/Sinks/osx/CoreAudioDevice.cpp
@@ -7,7 +7,7 @@
  */
 
 #include "CoreAudioDevice.h"
-#include "CoreAudioHelpers.h"
+#include "cores/AudioEngine/Sinks/darwin/CoreAudioHelpers.h"
 #include "CoreAudioChannelLayout.h"
 #include "CoreAudioHardware.h"
 #include "utils/log.h"
