+++ xbmc/cores/AudioEngine/Sinks/osx/CoreAudioHardware.cpp
@@ -8,7 +8,7 @@
 
 #include "CoreAudioHardware.h"
 
-#include "CoreAudioHelpers.h"
+#include "cores/AudioEngine/Sinks/darwin/CoreAudioHelpers.h"
 #include "utils/log.h"
 #include "platform/darwin/DarwinUtils.h"
