+++ xbmc/cores/AudioEngine/Sinks/osx/CoreAudioStream.cpp
@@ -9,7 +9,7 @@
 #include "CoreAudioStream.h"
 #include "CoreAudioDevice.h"
 
-#include "CoreAudioHelpers.h"
+#include "cores/AudioEngine/Sinks/darwin/CoreAudioHelpers.h"
 #include "utils/log.h"
 
 CCoreAudioStream::CCoreAudioStream() :
